fix for updator

+++ repos/profiles.rb
@@ -127,6 +127,7 @@ module Repos
       end
 
       def filtered profile
+        return if profile[:phone].blank?
         profile.delete(:phone) if profile[:phone][:visible] == false
         return if profile[:productions].blank?
         profile[:productions].each { |production|
